backend/requirements.txt
about-time==4.2.1
aiodns==3.6.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp_socks==0.10.2
aiosignal==1.4.0
alive-progress==3.3.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
arabic-reshaper==3.0.0
asgiref==3.11.1
asn1crypto==1.5.1
asttokens==3.0.1
async-timeout==5.0.1
attrs==25.4.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.9.0
bs4==0.0.2
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.7
click==8.3.2
cloudscraper==1.2.71
colorama==0.4.6
cryptography==46.0.6
cssselect2==0.9.0
curl_cffi==0.15.0
decorator==5.2.1
distro==1.9.0
executing==2.2.1
fastapi==0.135.3
Flask==3.1.3
frozenlist==1.8.0
future==1.0.0
future-annotations==1.0.0
google-auth==2.49.1
google-genai==1.70.0
graphemeu==0.7.2
h11==0.16.0
holehe==1.61
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ipython==9.12.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonpickle==4.1.1
lxml==5.4.0
maigret==0.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mock==5.2.0
multidict==6.7.1
networkx==2.8.8
oscrypto==1.3.0
outcome==1.3.0.post0
parso==0.8.6
pexpect==4.9.0
pillow==12.2.0
platformdirs==4.9.4
prompt_toolkit==3.0.52
propcache==0.4.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycares==4.11.0
pycountry==24.6.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
pyHanko==0.34.1
pyhanko-certvalidator==0.30.2
pyparsing==3.3.2
pypdf==6.9.2
PyPDF2==3.0.1
PySocks==1.7.1
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-engineio==4.13.1
python-multipart==0.0.24
python-socketio==5.16.1
python-socks==2.8.1
pyvis==0.3.2
PyYAML==6.0.3
reportlab==4.4.10
requests==2.33.1
requests-futures==1.0.2
requests-toolbelt==1.0.0
rich==14.3.3
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
socid-extractor==0.0.27
sortedcontainers==2.4.0
soupsieve==2.8.3
stack-data==0.6.3
starlette==1.0.0
stem==1.8.2
svglib==1.5.1
tenacity==9.1.4
termcolor==3.3.0
tinycss2==1.5.1
tokenize_rt==6.2.0
torrequest==0.1.0
tqdm==4.67.3
traitlets==5.14.3
trio==0.33.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uritools==6.0.1
urllib3==2.6.3
uvicorn==0.44.0
wcwidth==0.6.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.8
wsproto==1.3.2
xhtml2pdf==0.2.17
XMind==1.2.0
yarl==1.23.0
